🚚 Renamed JS files
Renamed JS files to avoid the aditional index.js file
